// src/app/enums/condicion.ts
export enum CondicionFiscalCodigo {
  CF = 'CF',     // Consumidor final
  EX = 'EX',     // Exento
  FM = 'FM',     // Resp. Inscp. Fac. M
  GC = 'GC',     // Gran contribuyente
  RI = 'RI',     // Responsable inscripto
  RMT = 'RMT',   // Responsable monotributo
  RNI = 'RNI',   // No inscripto
  SSF = 'SSF',   // Sin situación fiscal
  CDE = 'CDE'    // Cliente del exterior
}