// src/app/interfaces/provincia.ts
export interface Provincia{

  id?:number;
  descripcion:string;
  provinciaRiv:EProvincia;

}


export enum EProvincia {
  BUENOS_AIRES = "BUENOS_AIRES",
  CAPITAL_FEDERAL = "CAPITAL_FEDERAL",
  CATAMARCA = "CATAMARCA",
  CORDOBA = "CORDOBA",
  CORRIENTES = "CORRIENTES",
  CHACO = "CHACO",
  CHUBUT = "CHUBUT",
  ENTRE_RIOS = "ENTRE_RIOS",
  FORMOSA = "FORMOSA",
  TIERRA_DEL_FUEGO = "TIERRA_DEL_FUEGO",
  JUJUY = "JUJUY",
  LA_PAMPA = "LA_PAMPA",
  LA_RIOJA = "LA_RIOJA",
  MENDOZA = "MENDOZA",
  MISIONES = "MISIONES",
  NEUQUEN = "NEUQUEN",
  RIO_NEGRO = "RIO_NEGRO",
  SALTA = "SALTA",
  SAN_JUAN = "SAN_JUAN",
  SAN_LUIS = "SAN_LUIS",
  SANTA_CRUZ = "SANTA_CRUZ",
  SANTA_FE = "SANTA_FE",
  SANTIAGO_DEL_ESTERO = "SANTIAGO_DEL_ESTERO",
  TUCUMAN = "TUCUMAN"
}